fix(migrate recurring payment endpoint) the recurring payment endpoint has been transferred from insights controller to transaction controller
