+++ src/main/kotlin/com/nbk/insights/controller/TransactionController.kt
@@ -2,7 +2,9 @@ package com.nbk.insights.controller
 
 import com.nbk.insights.dto.TransactionResponse
 import com.nbk.insights.repository.UserRepository
+import com.nbk.insights.service.RecurringPaymentService
 import com.nbk.insights.service.TransactionsService
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.core.userdetails.UsernameNotFoundException
@@ -17,7 +19,8 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 class TransactionController(
     private val transactionsService: TransactionsService,
-    private val userRepository: UserRepository
+    private val userRepository: UserRepository,
+    private val recurringPaymentService: RecurringPaymentService
 ) {
 
     @GetMapping("/user/transactions")
@@ -42,6 +45,18 @@ class TransactionController(
             ?: throw UsernameNotFoundException("User not found with email: $username")
 
         return transactionsService.fetchAccountTransactions(accountId, user.id, mccId, period, category, year, month)
+    }
+
+    @GetMapping("/account/recurring-payments/{accountId}")
+    fun detectRecurringPaymentsForAccount(
+        @PathVariable accountId: Long
+    ): ResponseEntity<Any> {
+        val username = SecurityContextHolder.getContext().authentication.name
+        val user = userRepository.findByUsername(username)
+            ?: return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(mapOf("error" to "user not found"))
+
+        return recurringPaymentService.detectRecurringPaymentsDynamic(accountId, user.id)
+
     }
 
 }
